+++ Cargo.toml
@@ -7,7 +7,6 @@ edition = "2021"
 [workspace]
 members = [
   "state-reconstruct-fetcher",
-  "state-reconstruct-fetcher/blobscan-client",
   "state-reconstruct-storage",
 ]
 
@@ -15,7 +14,6 @@ members = [
 async-trait = "0.1.74"
 bincode = "1"
 blake2 = "0.10.6"
-blobscan-client = { path = "./state-reconstruct-fetcher/blobscan-client" }
 chrono = "0.4.31"
 clap = { version = "4.4.7", features = ["derive", "env"] }
 ethers = "1.0.2"
