+++ state-reconstruct-fetcher/Cargo.toml
@@ -6,7 +6,6 @@ edition = "2021"
 [dependencies]
 bincode = "1.3.3"
 blake2 = "0.10.6"
-blobscan-client = { path = "./blobscan-client" }
 ethers = "1.0.2"
 eyre = "0.6.8"
 indexmap = { version = "2.0.2", features = ["serde"] }
