--- state-reconstruct-fetcher/blobscan-client/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-name = "blobscan-client"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-eyre = "0.6.8"
-hex = "0.4.3"
-reqwest = "0.11.24"
-serde = { version = "1.0.189", features = ["derive"] }
-serde_json = { version = "1.0.107", features = ["std"] }
-thiserror = "1.0.50"
-tokio = { version = "1.33.0", features = ["signal"] }
-tracing = "0.1.40"
