--- state-reconstruct-fetcher/blobscan-client/src/lib.rs
@@ -1,4 +1,0 @@
-mod blob_support;
-pub use crate::blob_support::{BlobResponseFormatError, BlobSupport};
-mod scraping_support;
-pub use crate::scraping_support::ScrapingSupport;
