+++ state-reconstruct-fetcher/src/l1_fetcher.rs
@@ -1,6 +1,5 @@
 use std::{cmp, fs::File, future::Future, sync::Arc};
 
-use blobscan_client::{BlobSupport, ScrapingSupport};
 use ethers::{
     abi::{Contract, Function},
     prelude::*,
@@ -16,7 +15,6 @@ use tokio::{
 use tokio_util::sync::CancellationToken;
 
 use crate::{
-    api_support::ApiSupport,
     blob_http_client::BlobHttpClient,
     constants::ethereum::{BLOB_BLOCK, BLOCK_STEP, BOOJUM_BLOCK, GENESIS_BLOCK, ZK_SYNC_ADDR},
     metrics::L1Metrics,
@@ -468,7 +466,7 @@ impl L1Fetcher {
     ) -> Result<tokio::task::JoinHandle<Option<u64>>> {
         let metrics = self.metrics.clone();
         let contracts = self.contracts.clone();
-        let client = BlobHttpClient::new(make_support(self.config.blobs_url.clone()))?;
+        let client = BlobHttpClient::new(self.config.blobs_url.clone())?;
         Ok(tokio::spawn({
             async move {
                 let mut boojum_mode = false;
@@ -507,8 +505,8 @@ impl L1Fetcher {
                                     }
                                     sleep(Duration::from_secs(LONG_POLLING_INTERVAL_S)).await;
                                 }
-                                ParseError::BlobFormatError(inner) => {
-                                    tracing::error!("Cannot parse {}: {}", inner.0, inner.1);
+                                ParseError::BlobFormatError(data, inner) => {
+                                    tracing::error!("Cannot parse {}: {}", data, inner);
                                     cancellation_token.cancel();
                                     return last_block_number_processed;
                                 }
@@ -561,14 +559,6 @@ impl L1Fetcher {
             }
         }
         Err(err.into())
-    }
-}
-
-fn make_support(url: String) -> Box<dyn BlobSupport + Send + Sync> {
-    if url.starts_with("https://blobscan.com") {
-        Box::<ScrapingSupport>::default()
-    } else {
-        Box::new(ApiSupport::new(url))
     }
 }
