+++ state-reconstruct-fetcher/src/lib.rs
@@ -1,7 +1,6 @@
 #![feature(array_chunks)]
 #![feature(iter_next_chunk)]
 
-pub mod api_support;
 pub mod blob_http_client;
 pub mod constants;
 pub mod l1_fetcher;
